myapp-backend/requirements.txt
# FastAPI 및 ASGI 서버
fastapi==0.115.9
uvicorn[standard]==0.34.2
gunicorn==23.0.0

# Prisma ORM
prisma==0.15.0

# 비밀번호 해시 처리
bcrypt==5.0.0

# 데이터 검증
pydantic==2.12.3
pydantic-settings==2.10.1

# 환경 변수 관리
python-dotenv==1.1.0

# HTTP 클라이언트 (Prisma 의존성)
httpx==0.28.1

# JWT 인증
PyJWT==2.10.1
python-jose[cryptography]==3.3.0

# 기타 의존성
typing-extensions==4.15.0

# SQLAlchemy (database.py에서 사용)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
